refactor: rename query and mutation hooks with explicit suffixes
