+++ src/features/tracks/hooks/use-genres.ts
@@ -1,8 +1,8 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { api, ApiError } from "../api/api";
+import { api, ApiError } from "@/features/tracks/api";
 
-export const useGenres = () => {
+export const useGenresQuery = () => {
   return useQuery<string[], ApiError>({
     queryKey: ["genres"],
     queryFn: api.getGenres,
